Move typescript package into subdirectory
