+++ .github/workflows/ci.yml
@@ -138,12 +138,12 @@ jobs:
 
       - run: yarn install --immutable
       - run: yarn lint:ci
-      - run: yarn typecheck
+      - run: yarn build
       - run: yarn test
-      - run: yarn update-generated-files
 
       - name: Ensure generated files are up to date
         run: |
+          yarn update-generated-files
           git add --all
           changes=$(git status --porcelain)
           if [ -n "$changes" ]; then
@@ -170,7 +170,7 @@ jobs:
 
       - name: Publish to NPM
         if: startsWith(github.ref, 'refs/tags/typescript/v')
-        run: yarn publish --access public
+        run: yarn npm publish --access public
         env:
           NODE_AUTH_TOKEN: ${{ secrets.NPM_PUBLISH_TOKEN }}
