--- jsonschema/index.ts
@@ -1,3 +1,0 @@
-// This file exists to allow packages to import via `import { Foo } from "@foxglove/schemas/jsonschema"`.
-
-export * from "../schemas/jsonschema";
