+++ package.json
@@ -1,25 +1,12 @@
 {
-  "name": "@foxglove/schemas",
-  "version": "1.6.6",
-  "description": "Foxglove-defined message schemas for ROS, Protobuf, FlatBuffers, OMG IDL, and JSON",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/foxglove/schemas.git"
-  },
-  "author": {
-    "name": "Foxglove Technologies Inc",
-    "email": "support@foxglove.dev"
-  },
-  "engines": {
-    "node": ">=16"
-  },
+  "name": "foxglove-sdk",
+  "private": true,
   "packageManager": "yarn@4.6.0",
-  "main": "schemas/typescript/index.js",
-  "typings": "schemas/typescript/index.d.ts",
+  "workspaces": [
+    "typescript/*"
+  ],
   "scripts": {
-    "prepack": "yarn update-generated-files && tsc -p tsconfig.json",
-    "typecheck": "tsc -p tsconfig.json --noEmit",
+    "build": "tsc -b",
     "lint:ci": "eslint --report-unused-disable-directives .",
     "lint": "eslint --report-unused-disable-directives --fix .",
     "test": "jest",
@@ -51,9 +38,5 @@
     "ts-jest": "29.2.5",
     "ts-node": "^10.9.2",
     "typescript": "5.7.3"
-  },
-  "dependencies": {
-    "@foxglove/rosmsg-msgs-common": "^3.0.0",
-    "tslib": "^2.8.1"
   }
 }
