+++ typescript/schemas/README.md
@@ -1,0 +1,3 @@
+# @foxglove/schemas
+
+This package contains TypeScript type definitions for the Foxglove schemas.
