+++ typescript/schemas/package.json
@@ -1,0 +1,39 @@
+{
+  "name": "@foxglove/schemas",
+  "version": "1.6.6",
+  "description": "Foxglove-defined message schemas for ROS, Protobuf, FlatBuffers, OMG IDL, and JSON",
+  "license": "MIT",
+  "author": {
+    "name": "Foxglove",
+    "email": "support@foxglove.dev",
+    "url": "https://foxglove.dev/"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/foxglove/foxglove-sdk.git",
+    "directory": "typescript/schemas"
+  },
+  "main": "./dist/typescript/schemas/src/index.js",
+  "exports": {
+    ".": "./dist/typescript/schemas/src/index.js",
+    "./internal": "./dist/typescript/schemas/src/internal.js",
+    "./jsonschema": "./dist/typescript/schemas/src/jsonschema.js"
+  },
+  "files": [
+    "dist"
+  ],
+  "packageManager": "yarn@4.6.0",
+  "scripts": {
+    "build": "tsc -b",
+    "clean": "rimraf dist",
+    "prepack": "yarn workspace foxglove-sdk update-generated-files && yarn build"
+  },
+  "dependencies": {
+    "tslib": "^2"
+  },
+  "devDependencies": {
+    "@foxglove/rosmsg-msgs-common": "^3.0.0",
+    "rimraf": "^6.0.1",
+    "typescript": "5.7.3"
+  }
+}
