+++ typescript/schemas/src/internal/generateFlatbufferSchema.ts
