+++ typescript/schemas/src/internal/generateRos.ts
