+++ typescript/schemas/src/internal/generateTypeScript.test.ts
@@ -115,7 +115,7 @@ describe("generateTypeScript", () => {
   });
 
   it("generates parseable .ts files", async () => {
-    await expect(import("../schemas/typescript")).resolves.not.toThrow();
+    await expect(import("../../../../schemas/typescript")).resolves.not.toThrow();
   });
 
   it("can include TypedArray types", () => {
