+++ typescript/schemas/src/jsonschema.ts
@@ -1,0 +1,1 @@
+export * from "../../../schemas/jsonschema";
